Allow matching a driver for the interface link
Add a property to allow matching a driver for the persistent network
naming configuration. Also, sets the MACAddressPolicy as `persistent`
for all the configured links. This will do nothing on ethernet
adaptors that already have a real persistent MAC Address, but will
ensure that the cheap ones will not get a different MAC Address on each
system boot (for the current system).

Signed-off-by: Filipe Utzig <filipe@gringolito.com>