Network configuration improvements (#5)
* Fix the networking configuration handler

Hosts don't need to reboot for applying network configurations. A simple
restart on the networking service can handle it.

Signed-off-by: Filipe Utzig <filipe@gringolito.com>

* Add support for Proxmox SDN on the interfaces file

In order for the Proxmox SDN configuration to work we need to include
all the files under the `/etc/network/interfaces.d/*`.

Source: https://pve.proxmox.com/pve-docs/chapter-pvesdn.html#_sdn_core

Signed-off-by: Filipe Utzig <filipe@gringolito.com>

* Allow configuring non-bridged interfaces

Allow configuration non-bridged interfaces as VLANs or regular
interfaces on the `/etc/network/interfaces` file.

Signed-off-by: Filipe Utzig <filipe@gringolito.com>

* Allow matching a driver for the interface link

Add a property to allow matching a driver for the persistent network
naming configuration. Also, sets the MACAddressPolicy as `persistent`
for all the configured links. This will do nothing on ethernet
adaptors that already have a real persistent MAC Address, but will
ensure that the cheap ones will not get a different MAC Address on each
system boot (for the current system).

Signed-off-by: Filipe Utzig <filipe@gringolito.com>

---------

Signed-off-by: Filipe Utzig <filipe@gringolito.com>